+++ utils-test-app/utils_test_app/tests/test_testing.py
@@ -2,8 +2,14 @@ from django.test import TestCase
 
 import requests
 from django.contrib.auth.models import User
+from allianceauth.authentication.models import EveCharacter
 
-from app_utils.testing import NoSocketsTestCase, SocketAccessError, generate_invalid_pk
+from app_utils.testing import (
+    NoSocketsTestCase,
+    SocketAccessError,
+    generate_invalid_pk,
+    create_user_from_evecharacter,
+)
 
 
 class TestNoSocketsTestCase(NoSocketsTestCase):
@@ -20,3 +26,35 @@ class TestGenerateInvalidPk(TestCase):
         invalid_pk = generate_invalid_pk(User)
         with self.assertRaises(User.DoesNotExist):
             User.objects.get(pk=invalid_pk)
+
+
+class TestCreateUserFromEveCharacter(TestCase):
+    @classmethod
+    def setUpClass(cls) -> None:
+        super().setUpClass()
+        cls.character = EveCharacter.objects.create(
+            character_id=1001,
+            character_name="Bruce Wayne",
+            corporation_id=2001,
+            corporation_name="Wayne Tech",
+            corporation_ticker="WYT",
+        )
+
+    def test_should_create_basic_user(self):
+        # when
+        user, character_ownership = create_user_from_evecharacter(1001)
+        # then
+        self.assertEqual(user.username, "Bruce_Wayne")
+        self.assertEqual(character_ownership.character, self.character)
+        self.assertEqual(character_ownership.user, user)
+
+    def test_should_create_user_with_given_scope(self):
+        # when
+        user, character_ownership = create_user_from_evecharacter(
+            1001, scopes=["dummy_scope"]
+        )
+        # then
+        self.assertEqual(user.username, "Bruce_Wayne")
+        self.assertEqual(character_ownership.character, self.character)
+        self.assertEqual(character_ownership.user, user)
+        self.assertTrue(user.token_set.filter(scopes__name="dummy_scope").exists())