+++ CHANGELOG.md
@@ -9,9 +9,24 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
+- AttrDict: Access dict like an object
+
+### Changed
+
+- Performance improvement for django.users_with_permission
+
+### Fixed
+
+- Create TZ aware timestamps in fake ESI response
+
+## [1.1.0] - 2021-03-19
+
+### Added
+
 - testing.create_user_from_evecharacter
 - esi_testing module with tools for testing with django-esi
-- testing:  response_text, json_response_to_python, json_response_to_dict, multi_assert_in, multi_assert_not_in
+- testing: response_text, json_response_to_python, json_response_to_dict, multi_assert_in, multi_assert_not_in
+
 
 ## [1.0.2] - 2021-03-13